migrate DI solution (GSI-440)
Switched to DI based on basic Python constructs like context managers.

Upgrades to pydantic v2.

Switches to docker-outside-of-docker for the devcontainer environment.

Bumps version to 1.0.0.

The tests do not use the standard JointFixture setup, however, to keep the PR small, this is not changed. May be reconsidered in a separate PR.